--- src/main/java/org/tanzu/mcpclient/util/McpDiscoveryService.java
@@ -1,123 +1,0 @@
-package org.tanzu.mcpclient.util;
-
-import io.pivotal.cfenv.boot.genai.GenaiLocator;
-import io.pivotal.cfenv.core.CfCredentials;
-import io.pivotal.cfenv.core.CfEnv;
-import io.pivotal.cfenv.core.CfService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.lang.Nullable;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-/**
- * Service for discovering and managing Model Context Protocol (MCP) service connections.
- * Handles MCP service discovery from both GenaiLocator and Cloud Foundry service bindings.
- */
-@Service
-public class McpDiscoveryService {
-    private static final Logger logger = LoggerFactory.getLogger(McpDiscoveryService.class);
-
-    // Constants
-    public static final String MCP_SERVICE_URL = "mcpServiceURL";
-
-    private final CfEnv cfEnv;
-    private final GenaiLocator genaiLocator; // Optional - may be null
-
-    /**
-     * Constructor with optional GenaiLocator injection.
-     * GenaiLocator is only available when GenaiLocatorAutoConfiguration is active
-     * (i.e., when genai.locator.config-url property is set by CfGenaiProcessor).
-     */
-    public McpDiscoveryService(@Nullable GenaiLocator genaiLocator) {
-        this.cfEnv = new CfEnv();
-        this.genaiLocator = genaiLocator;
-
-        if (genaiLocator != null) {
-            logger.debug("GenaiLocator bean detected - will check for dynamic MCP service discovery");
-        } else {
-            logger.debug("No GenaiLocator bean available - using CF service-based MCP discovery only");
-        }
-    }
-
-    /**
-     * Gets MCP server URLs from GenaiLocator if available, otherwise returns empty list.
-     * This is the new approach for MCP service discovery.
-     */
-    public List<String> getMcpServiceUrlsFromLocator() {
-        if (genaiLocator == null) {
-            return List.of();
-        }
-        try {
-            return genaiLocator.getMcpServers().stream()
-                    .map(GenaiLocator.McpConnectivity::url)
-                    .collect(Collectors.toList());
-        } catch (Exception e) {
-            logger.debug("Error getting MCP service URLs from GenaiLocator: {}", e.getMessage());
-            return List.of();
-        }
-    }
-
-    /**
-     * Gets the names of MCP services from Cloud Foundry service bindings.
-     * This is the legacy approach for MCP service discovery.
-     */
-    public List<String> getMcpServiceNames() {
-        try {
-            return cfEnv.findAllServices().stream()
-                    .filter(this::hasMcpServiceUrl)
-                    .map(CfService::getName)
-                    .collect(Collectors.toList());
-        } catch (Exception e) {
-            logger.warn("Error getting MCP service names: {}", e.getMessage());
-            return List.of();
-        }
-    }
-
-    /**
-     * Gets the URLs of MCP services from Cloud Foundry service bindings.
-     * This is the legacy approach - new approach uses GenaiLocator.
-     */
-    public List<String> getMcpServiceUrls() {
-        try {
-            return cfEnv.findAllServices().stream()
-                    .filter(this::hasMcpServiceUrl)
-                    .map(service -> service.getCredentials().getString(MCP_SERVICE_URL))
-                    .collect(Collectors.toList());
-        } catch (Exception e) {
-            logger.warn("Error getting MCP service URLs: {}", e.getMessage());
-            return List.of();
-        }
-    }
-
-    /**
-     * Gets all MCP service URLs from both sources (GenaiLocator and CF services).
-     * This ensures compatibility with both old and new approaches.
-     */
-    public List<String> getAllMcpServiceUrls() {
-        List<String> locatorUrls = getMcpServiceUrlsFromLocator();
-        List<String> cfUrls = getMcpServiceUrls();
-
-        // Combine both sources, preferring GenaiLocator if available
-        if (!locatorUrls.isEmpty()) {
-            logger.debug("Using MCP service URLs from GenaiLocator: {}", locatorUrls);
-            return locatorUrls;
-        } else if (!cfUrls.isEmpty()) {
-            logger.debug("Using MCP service URLs from CF services: {}", cfUrls);
-            return cfUrls;
-        } else {
-            logger.debug("No MCP service URLs found from any source");
-            return List.of();
-        }
-    }
-
-    /**
-     * Checks if a Cloud Foundry service has an MCP service URL configured.
-     */
-    public boolean hasMcpServiceUrl(CfService service) {
-        CfCredentials credentials = service.getCredentials();
-        return credentials != null && credentials.getString(MCP_SERVICE_URL) != null;
-    }
-}
