--- src/main/java/org/tanzu/mcpclient/util/ModelSource.java
@@ -1,16 +1,0 @@
-package org.tanzu.mcpclient.util;
-
-/**
- * Enum representing the source of model configuration.
- */
-public enum ModelSource {
-    /**
-     * Model configuration comes from GenaiLocator (Tanzu Platform 10.2+ format).
-     */
-    GENAI_LOCATOR,
-    
-    /**
-     * Model configuration comes from traditional Spring AI properties.
-     */
-    PROPERTIES
-}
