--- src/main/java/org/tanzu/mcpclient/util/PropertyBasedModelProvider.java
@@ -1,140 +1,0 @@
-package org.tanzu.mcpclient.util;
-
-import io.micrometer.observation.ObservationRegistry;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.ai.chat.model.ChatModel;
-import org.springframework.ai.document.MetadataMode;
-import org.springframework.ai.embedding.EmbeddingModel;
-import org.springframework.ai.model.tool.ToolCallingManager;
-import org.springframework.ai.openai.OpenAiChatModel;
-import org.springframework.ai.openai.OpenAiChatOptions;
-import org.springframework.ai.openai.OpenAiEmbeddingModel;
-import org.springframework.ai.openai.OpenAiEmbeddingOptions;
-import org.springframework.ai.openai.api.OpenAiApi;
-import org.springframework.retry.support.RetryTemplate;
-import org.springframework.stereotype.Component;
-
-import java.util.Optional;
-
-/**
- * ModelProvider implementation that provides models from traditional Spring AI properties.
- * This has lower priority than GenaiLocator but serves as a fallback for property-based configuration.
- */
-@Component
-public class PropertyBasedModelProvider implements ModelProvider {
-    
-    private static final Logger logger = LoggerFactory.getLogger(PropertyBasedModelProvider.class);
-    private static final int PRIORITY = 10; // Lower priority than GenaiLocator
-    
-    private final GenAIService genAIService;
-    private final OpenAiApi openAiApi;
-    private final RetryTemplate retryTemplate;
-    private final ObservationRegistry observationRegistry;
-    private final ToolCallingManager toolCallingManager;
-    
-    public PropertyBasedModelProvider(GenAIService genAIService, 
-                                    OpenAiApi openAiApi, 
-                                    RetryTemplate retryTemplate,
-                                    ObservationRegistry observationRegistry, 
-                                    ToolCallingManager toolCallingManager) {
-        this.genAIService = genAIService;
-        this.openAiApi = openAiApi;
-        this.retryTemplate = retryTemplate;
-        this.observationRegistry = observationRegistry;
-        this.toolCallingManager = toolCallingManager;
-    }
-    
-    @Override
-    public Optional<ChatModel> getChatModel() {
-        ModelConfig config = genAIService.getChatModelConfig();
-        
-        // Only provide model if it's from properties (not GenaiLocator)
-        if (config.isFromGenaiLocator()) {
-            logger.debug("Chat model config is from GenaiLocator - PropertyBasedModelProvider will not provide model");
-            return Optional.empty();
-        }
-        
-        try {
-            ChatModel chatModel = createPropertyBasedChatModel(config);
-            logger.debug("Successfully created property-based ChatModel with model='{}'", config.modelName());
-            return Optional.of(chatModel);
-        } catch (Exception e) {
-            logger.warn("Failed to create property-based ChatModel: {}", e.getMessage());
-            return Optional.empty();
-        }
-    }
-    
-    @Override
-    public Optional<EmbeddingModel> getEmbeddingModel() {
-        ModelConfig config = genAIService.getEmbeddingModelConfig();
-        
-        // Only provide model if it's from properties (not GenaiLocator)
-        if (config.isFromGenaiLocator()) {
-            logger.debug("Embedding model config is from GenaiLocator - PropertyBasedModelProvider will not provide model");
-            return Optional.empty();
-        }
-        
-        try {
-            EmbeddingModel embeddingModel = createPropertyBasedEmbeddingModel(config);
-            logger.debug("Successfully created property-based EmbeddingModel with model='{}'", config.modelName());
-            return Optional.of(embeddingModel);
-        } catch (Exception e) {
-            logger.warn("Failed to create property-based EmbeddingModel: {}", e.getMessage());
-            return Optional.empty();
-        }
-    }
-    
-    @Override
-    public int getPriority() {
-        return PRIORITY;
-    }
-    
-    @Override
-    public String getProviderName() {
-        return "PropertyBased";
-    }
-    
-    /**
-     * Creates a property-based ChatModel using traditional Spring AI configuration.
-     */
-    private ChatModel createPropertyBasedChatModel(ModelConfig config) {
-        String model = config.modelName();
-        
-        if (model == null || model.isEmpty()) {
-            logger.warn("No chat model configured in properties - creating non-functional ChatModel bean");
-            model = "no-model-configured";
-        } else {
-            logger.info("Creating property-based ChatModel with model='{}', apiKey exists={}",
-                    model, config.isValid());
-        }
-        
-        OpenAiChatOptions options = OpenAiChatOptions.builder()
-                .model(model)
-                .temperature(0.8)
-                .build();
-        
-        return new OpenAiChatModel(openAiApi, options, toolCallingManager, retryTemplate, observationRegistry);
-    }
-    
-    /**
-     * Creates a property-based EmbeddingModel using traditional Spring AI configuration.
-     */
-    private EmbeddingModel createPropertyBasedEmbeddingModel(ModelConfig config) {
-        String model = config.modelName();
-        
-        if (model == null || model.isEmpty()) {
-            logger.warn("No embedding model configured in properties - creating non-functional EmbeddingModel bean");
-            model = "no-model-configured";
-        } else {
-            logger.info("Creating property-based EmbeddingModel with model='{}', apiKey exists={}",
-                    model, config.isValid());
-        }
-        
-        OpenAiEmbeddingOptions options = OpenAiEmbeddingOptions.builder()
-                .model(model)
-                .build();
-        
-        return new OpenAiEmbeddingModel(openAiApi, MetadataMode.EMBED, options, retryTemplate);
-    }
-}
